feat: OpenAI API 호출 변경

+++ src/main/java/com/example/plantappbackend/repository/PlantRepository.java
@@ -10,10 +10,9 @@ import java.util.List;
 
 @Repository
 public interface PlantRepository extends JpaRepository<Plant, Integer> {
-    List<Plant> findByUserId(int userId); // 특정 사용자 ID로 식물 목록 조회
+    List<Plant> findByUserId(Long userId); // 특정 사용자 ID로 식물 목록 조회
 
     // 특정 닉네임으로 식물 조회 (Room의 @Query 스타일)
     @Query("SELECT p FROM Plant p WHERE p.nickname = :nickname")
     List<Plant> findByNickname(@Param("nickname") String nickname);
-
 }
